backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pymongo[srv]
google-genai
pydantic
